addressing reviews.

+++ beacon/validator/src/main/java/tech/pegasys/teku/validator/coordinator/performance/DefaultPerformanceTracker.java
@@ -289,7 +289,7 @@ public class DefaultPerformanceTracker implements PerformanceTracker {
     int numberOfProducedAttestations = producedAttestations.size();
     return producedAttestations.size() > 0
         ? new AttestationPerformance(
-            analyzedEpoch.intValue(),
+            analyzedEpoch,
             validatorTracker.getNumberOfValidatorsForEpoch(analyzedEpoch),
             numberOfProducedAttestations,
             (int) inclusionDistanceStatistics.getCount(),
@@ -299,8 +299,7 @@ public class DefaultPerformanceTracker implements PerformanceTracker {
             correctTargetCount,
             correctHeadBlockCount)
         : AttestationPerformance.empty(
-            analyzedEpoch.intValue(),
-            validatorTracker.getNumberOfValidatorsForEpoch(analyzedEpoch));
+            analyzedEpoch, validatorTracker.getNumberOfValidatorsForEpoch(analyzedEpoch));
   }
 
   private Set<BeaconBlock> getBlocksInEpochs(UInt64 startEpochInclusive, UInt64 endEpochExclusive) {
